+++ app/(chat)/api/rag/route.ts
@@ -1,15 +1,23 @@
-// app/api/test-rag/route.ts
+export const config = { runtime: 'nodejs' };
+
 import { NextResponse } from "next/server";
 import { tool } from "ai";
 import { Document, VectorStoreIndex } from "llamaindex";
 import { z } from "zod";
+import { promises as fs } from "fs";
+import { join } from "path";
+import { SimpleDirectoryReader } from "@llamaindex/readers/directory";
+import { PDFReader } from '@llamaindex/readers/pdf';
+import { QdrantVectorStore } from "@llamaindex/qdrant";
 
-// Create an index from your documents
-const document = new Document({
+const TEMP_DIR = "./tmp";
+
+// Create a sample document and index for testing
+const sampleDocument = new Document({
   text: "The author is Kenji, and Kenji studied computer science at Manchester",
   id_: "unique-id"
 });
-const index = await VectorStoreIndex.fromDocuments([document]);
+const sampleIndex = await VectorStoreIndex.fromDocuments([sampleDocument]);
 
 // Define the query tool
 export const queryTool = tool({
@@ -19,13 +27,12 @@ export const queryTool = tool({
     query: z.string().describe("The question or query about the document content"),
   }),
   execute: async ({ query }) => {
-    const queryEngine = index.asQueryEngine();
+    const queryEngine = sampleIndex.asQueryEngine();
     const response = await queryEngine.query({ query });
     return {
       answer: response.message.content,
       sources: response.sourceNodes
         ? response.sourceNodes.map((node) => {
-            // Cast the node to Document to access its content.
             const doc = node.node as unknown as Document;
             const content = (doc as any).text || "";
             return {
@@ -54,7 +61,75 @@ async function runRAGTest() {
   return { queryResponse };
 }
 
-// API GET handler that calls our test function
+// PDF Processing Functions
+async function processPDF(file: File, query: string | null): Promise<any> {
+  // Ensure temporary directory exists
+  await fs.mkdir(TEMP_DIR, { recursive: true });
+  
+  // Write the uploaded file to disk
+  const arrayBuffer = await file.arrayBuffer();
+  const buffer = Buffer.from(arrayBuffer);
+  const filePath = join(TEMP_DIR, file.name);
+  await fs.writeFile(filePath, buffer);
+  
+  // Use SimpleDirectoryReader with PDFReader
+  const reader = new SimpleDirectoryReader();
+  const docs = await reader.loadData({
+    directoryPath: TEMP_DIR,
+    fileExtToReader: {
+      pdf: new PDFReader(),
+    },
+  });
+  
+  // Get parsed text from all documents for response
+  const parsedText = docs.map(doc => doc.text).join("\n\n");
+  
+  // Create page-chunked documents for indexing
+  const chunkedDocuments = docs.flatMap((doc) => {
+    // Split by page markers if available in text
+    const pages = doc.text.split(/\f|\n---Page \d+---\n/);
+    
+    return pages.filter(page => page.trim().length > 0).map((pageText, idx) => {
+      return new Document({ 
+        text: pageText, 
+        id_: `${doc.id_}_page${idx + 1}`,
+        metadata: { 
+          ...doc.metadata,
+          page_number: idx + 1,
+          source: filePath,
+          filename: file.name
+        }
+      });
+    });
+  });
+  
+  // Connect to Qdrant vector store
+  const vectorStore = new QdrantVectorStore({
+    url: "http://localhost:6333",
+  });
+  
+  // Create vector index from chunked documents
+  const index = await VectorStoreIndex.fromDocuments(chunkedDocuments, {
+    vectorStore,
+  });
+  
+  // If query parameter is provided, perform a search
+  let queryResult = null;
+  if (query) {
+    const queryEngine = index.asQueryEngine();
+    queryResult = await queryEngine.query({
+      query,
+    });
+  }
+  
+  return {
+    parsedText,
+    indexId: chunkedDocuments.map(doc => doc.id_).join(","),
+    queryResult: queryResult?.toString()
+  };
+}
+
+// API GET handler - Test the RAG functionality with sample document
 export async function GET() {
   try {
     const results = await runRAGTest();
@@ -62,5 +137,34 @@ export async function GET() {
   } catch (error: any) {
     console.error("Error in GET:", error);
     return NextResponse.json({ success: false, error: error.message }, { status: 500 });
+  }
+}
+
+// API POST handler - Process PDF uploads and optional querying
+export async function POST(req: Request): Promise<Response> {
+  try {
+    const formData = await req.formData();
+    const file = formData.get("file") as File | null;
+    const query = formData.get("query") as string | null;
+    
+    if (!file) {
+      return new Response(
+        JSON.stringify({ message: "No file uploaded" }),
+        { status: 400, headers: { "Content-Type": "application/json" } }
+      );
+    }
+    
+    const result = await processPDF(file, query);
+    
+    return new Response(
+      JSON.stringify(result),
+      { status: 200, headers: { "Content-Type": "application/json" } }
+    );
+  } catch (error: unknown) {
+    const errorMessage = error instanceof Error ? error.message : "Error processing PDF";
+    return new Response(
+      JSON.stringify({ message: errorMessage }),
+      { status: 500, headers: { "Content-Type": "application/json" } }
+    );
   }
 }
